# docs/swagger.yaml
basePath: /api/v1
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API para consulta de dados do Ragnarok Online
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Latam RO Database API
  version: "1.0"
paths:
  /monsters:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de todos os monstros cadastrados
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Lista todos os monstros
      tags:
      - monsters
  /monsters/{id}:
    get:
      consumes:
      - application/json
      description: Retorna os detalhes de um monstro pelo ID
      parameters:
      - description: ID do monstro
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Busca um monstro específico
      tags:
      - monsters
  /monsters/type/{type}:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de monstros filtrados por tipo (fire, water,
        etc)
      parameters:
      - description: Tipo do monstro
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Busca monstros por tipo
      tags:
      - monsters
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"